requirements.txt
accelerate==1.0.1
aiohttp==3.10.11
anime_face_detector==0.0.9
beautifulsoup4==4.13.3
braceexpand==0.1.7
clip_retrieval==2.44.0
colorama==0.4.6
compel==2.0.3
consisid_eva_clip==1.0.2
controlnet_aux==0.0.9
d4rl==1.1
dadaptation==3.2
datasets==3.1.0
decord==0.6.0
deepspeed==0.16.5
easydict==1.13
einops==0.8.1
facexlib==0.3.0
fairscale==0.4.13
faiss==1.5.3
fastapi==0.115.12
flax==0.10.5
ftfy==6.3.1
GitPython==3.1.44
gym==0.26.2
huggingface_hub==0.30.1
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_metadata==8.0.0
imwatermark==0.0.2
insightface==0.7.3
intel_extension_for_pytorch==2.6.0
jax==0.5.3
jaxlib==0.5.3
jxlpy==0.9.5
k_diffusion==0.1.1.post1
kornia==0.8.0
librosa==0.11.0
ligo==1.0.9
lion_pytorch==0.2.3
lpips==0.1.4
matplotlib==3.7.5
modelscope==1.24.1
msgpack_python==0.5.6
muse==0.1
neural_compressor==3.3.1
note_seq==0.0.5
numpy==1.24.1
onnx==1.17.0
onnx_graphsurgeon==0.5.7
opencv_python==4.11.0.86
optax==0.2.4
optimum==1.24.0
packaging==24.2
pandas==2.0.3
parameterized==0.9.0
paths==0.0.1a1
Pillow==11.1.0
polygraphy==0.49.20
prodigyopt==1.1.2
pycuda==2025.1
pydantic==2.11.2
PyGithub==2.6.1
pytest==8.3.5
PyYAML==6.0.2
PyYAML==6.0.2
ram==0.1
regex==2024.11.6
Requests==2.32.3
requests_mock==1.12.1
rich==14.0.0
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
setuptools==75.3.0
skimage==0.0
slack_sdk==3.35.0
structlog==25.2.0
t5x==0.0.0
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow==2.19.0
tensorrt==10.9.0.34
termcolor==3.0.1
timm==1.0.15
toml==0.10.2
torch==2.4.1+cu118
torch_npu==2.4.0.post2
torch_xla==2.6.1
torchsde==0.2.6
torchvision==0.19.1+cu118
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.1
uvicorn==0.34.0
voluptuous==0.14.2
wandb==0.19.9
webdataset==0.2.111
xformers==0.0.29.post3
